feat: derived streams, plugin-scoped camera config, and IMuxStream rename
- introduce IPipeline, DerivedStreamPipeline, and StreamReconciler to
  manage derived-stream lifecycle alongside camera pipelines
- rename IVideoStream to IMuxStream and VideoStreamFanOut to
  MuxStreamFanOut; rename IVideoAnalyzer to IDataStreamAnalyzer
- add IPluginCameraSettings and IPluginStreamSettings, plus
  CoreCameraSettings, CoreStreamSettings, and CameraConfigService for
  plugin-scoped configuration
- extract sqlite migrations into MigrationScripts via DbUp and add
  the 0002_streams_derived migration
- update Onvif provider to surface plugin-scoped settings and rebuild
  the camera edit and storage views around the new config model
- expose camera-config endpoints in ApiRoutes and ApiExtensions
- update streaming, recording, and integration tests to the new
  contracts and rename

+++ src/Shared.Models/CameraConfiguration.cs
@@ -17,7 +17,7 @@ public sealed class StreamProfile
   public required string FormatId { get; init; }
   public string? Codec { get; init; }
   public string? Resolution { get; init; }
-  public int? Fps { get; init; }
+  public decimal? Fps { get; init; }
   public int? Bitrate { get; init; }
   public required string Uri { get; init; }
 }
